FEAT: Define the eventUI class that eventsUI will inherit
