+++ Assets/Scripts/UI/EventUI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d05f2f3866198db46b7afa0c93ca64bb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
