+++ Assets/Scripts/UI/EventUI/BlackSmithUI.cs
@@ -3,15 +3,14 @@ using Items;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class BlackSmithUI : MonoBehaviour
+public class BlackSmithUI : EventUI
 {
     private GameObject[] _options;
-    private GameObject _eventUIButton;
 
-    private void Start()
+    protected override void Start()
     {
+        base.Start();
         Transform childTransform = transform.Find("Options");
-        _eventUIButton = GameObject.Find("EventUIButton");
         Button[] childrenTransforms = childTransform.GetComponentsInChildren<Button>();
         _options = new GameObject[childrenTransforms.Length];
         for (int i = 0; i < childrenTransforms.Length; i++)
@@ -49,14 +48,4 @@ public class BlackSmithUI : MonoBehaviour
         }
     }
 
-    public void DisplayUI()
-    {
-        gameObject.SetActive(true);
-        _eventUIButton.gameObject.SetActive(false);
-    }
-    public void CloseUI()
-    {
-        gameObject.SetActive(false);
-        _eventUIButton.gameObject.SetActive(true);
-    }
 }
