+++ Assets/Scripts/UI/EventUI/EventUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2ae8a401cb091c64ca962dbd03bcc756
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
